crm-backend/requirements.txt
Django>=6.0.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
django-storages>=1.13.0
boto3>=1.26.0
django-filter>=23.2
Pillow>=10.0.0
django-cors-headers>=4.0.0
